bugfix: multiple definition of boost::json + refactoring of server integer types

+++ src/src/server.cpp
@@ -1,13 +1,14 @@
 #include "server.hpp"
 
 App::Server::Server() {
-    m_port = App::ConfigManager::get().getPort();
-    m_maxConnections = App::ConfigManager::get().getMaxConnections();
+    m_port = static_cast<uint_least16_t>(App::ConfigManager::get().getPort());
+    m_maxConnections = static_cast<uint_least16_t>(App::ConfigManager::get().getMaxConnections());
 
     m_activeConnections = 0;
 }
 
 awaitable<void> App::Server::listen() {
+    /*
     auto executor = co_await boost::asio::this_coro::executor;
     tcp::acceptor acceptor(executor, {tcp::v4(), m_port});
 
@@ -17,6 +18,7 @@ awaitable<void> App::Server::listen() {
         tcp::socket connection = co_await acceptor.async_accept();
         co_spawn(executor, write(std::move(connection)), boost::asio::detached);
     }
+    */
 }
 
 awaitable<void> App::Server::write(tcp::socket socket)
